Combine tests into one suite
